feat: Single reporter for VRF & RR
